Brand new error from enode

+++ src/main.rs
@@ -30,16 +30,16 @@ async fn main() -> anyhow::Result<()> {
     let mut auth_message = message::AuthMsgV4::default().rlp_bytes().to_vec();
     // append some zeros to make message distinguishable from non EIP-8 (required by eth)
     auth_message.extend_from_slice(&[0u8; 150]);
-    let auth_message_size: usize = auth_message.len() + ECIES_OVERHEAD;
+    let auth_message_size: u16 = (auth_message.len() + ECIES_OVERHEAD) as u16;
 
     let auth_encrypted = ecies::encrypt(
         auth_message.as_ref(),
         &node.public_key,
         &[],
-        &(auth_message_size as u16).to_be_bytes(),
+        &auth_message_size.to_be_bytes(),
     )?;
 
-    conn.write_u16(auth_message_size as u16).await?;
+    conn.write_u16(auth_message_size).await?;
     conn.write_all(&auth_encrypted).await?;
     conn.flush().await?;
     let res = conn.read_u16().await?;
